Move name merging function to `utils`
